Refactor Headers
