--- src/components/Head/Head.module.css
@@ -1,11 +1,0 @@
-/* .head {
-  position: absolute;
-  display: flex;
-  background-color: white;
-  box-shadow: 0 0 20px rgba(0, 0, 0, 0.3);
-
-  img {
-    height: 100%;
-    width: 100%;
-  }
-} */
