+++ src/components/Head/Head.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Foot } from '../Foot/Foot';
+import { Head } from './Head';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Foot />, div);
+  ReactDOM.render(<Head />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
